+++ common/src/main/java/dev/latvian/kubejs/server/KubeJSServerEventHandler.java
@@ -9,8 +9,8 @@ import dev.latvian.kubejs.player.SimplePlayerEventJS;
 import dev.latvian.kubejs.script.ScriptType;
 import dev.latvian.kubejs.util.ConsoleJS;
 import dev.latvian.kubejs.world.AttachWorldDataEvent;
-import dev.latvian.kubejs.world.ServerWorldJS;
-import dev.latvian.kubejs.world.SimpleWorldEventJS;
+import dev.latvian.kubejs.world.events.ServerWorldJS;
+import dev.latvian.kubejs.world.events.SimpleWorldEventJS;
 import dev.latvian.kubejs.world.WorldJS;
 import dev.latvian.mods.rhino.RhinoException;
 import me.shedaniel.architectury.event.events.CommandPerformEvent;
