+++ common/src/main/java/dev/latvian/kubejs/world/KubeJSWorldEventHandler.java
@@ -3,7 +3,10 @@ package dev.latvian.kubejs.world;
 import dev.latvian.kubejs.KubeJSEvents;
 import dev.latvian.kubejs.script.ScriptType;
 import dev.latvian.kubejs.server.ServerJS;
+import dev.latvian.kubejs.world.events.ExplosionEventJS;
 import dev.latvian.kubejs.world.events.LightningStrikeEventJS;
+import dev.latvian.kubejs.world.events.ServerWorldJS;
+import dev.latvian.kubejs.world.events.SimpleWorldEventJS;
 import me.shedaniel.architectury.event.events.ExplosionEvent;
 import me.shedaniel.architectury.event.events.LifecycleEvent;
 import me.shedaniel.architectury.event.events.LightningEvent;
